一些小更改和整合

+++ src/components/home.vue
@@ -1,63 +1,70 @@
 <template>
-  <div class="home-container">
-    <el-container class="home-container">
-      <!-- 头部区域 -->
-      <el-header>
-        <div>
-          <img src="../assets/heima.png" alt="vue">
-          <span>后台管理系统</span>
-        </div>
-        <el-button type="danger" @click="logout">退出登录</el-button>
-      </el-header>
+    <div class="home-container">
 
-      <!-- 页面主体区域 -->
-      <el-container>
+        <!-- 盒子一号 -->
+        <el-container class="home-container">
 
-        <!-- 侧边栏 -->
-        <el-aside :width="isCollapse ? '64px':'200px'">
-          <div class="toggle-button" @click="toggleCollapse">|||</div>
-          <!-- 侧边栏菜单区 -->
-          <el-menu class="el-menu-vertical-demo" background-color="RGB(252,234,241)" 
-          text-color="RGB(46,38,56)"
-            active-text-color="rgb(58, 178, 248)" :unique-opened="true" :collapse="isCollapse" :collapse-transition="false"
-            :router="true" :default-active="activePath">
-            <!-- submenu一级菜单 -->
-            <el-submenu :index="item.path" v-for="item in menulist" :key="item.id">
-              <!-- 一级菜单的内容渲染区域 -->
-              <template slot="title">
-                <!-- icon图标 -->
-                <i :class="iconObj[item.id]"></i>
-                <!-- 菜单文本 -->
-                <span>{{item.authName}}</span>
-              </template>
+            <!-- 头部 -->
+            <el-header>
+                <div>
+                    <img src="../assets/heima.png" alt="vue">
+                    <span>后台管理系统</span>
+                </div>
+                <el-button type="danger" size="small" round @click="logout">退出登录</el-button>
+            </el-header>
 
-<!-- 二级菜单 -->
-<el-menu-item :index="subItem.path" v-for="subItem in item.children" :key="subItem.id" @click="saveNavState(subItem.path)">
-    <template slot="title">
-                  <!-- icon图标 -->
-                  <i class="el-icon-menu"></i>
-                  <!-- 菜单文本 -->
-                  <span>{{subItem.authName}}</span>
-                </template>
-</el-menu-item>
-</el-submenu>
-</el-menu>
-</el-aside>
-<el-container>
+            <!-- 盒子二号 页面主体区域 -->
+            <el-container>
 
-    <!-- 内容主体区 -->
-    <el-main>
-        <router-view></router-view>
-    </el-main>
+                <!-- 侧边栏 -->
+                <el-aside :width="isCollapse ? '64px':'200px'">
+                    <div class="toggle-button" @click="toggleCollapse">|||</div>
+                    <!-- 侧边栏菜单区 -->
+                    <el-menu 
+                    class="el-menu-vertical-demo" 
+                    background-color="RGB(252,234,241)" 
+                    text-color="RGB(46,38,56)"
+                        active-text-color="rgb(58, 178, 248)" 
+                        :unique-opened="true" :collapse="isCollapse" 
+                        :collapse-transition="false"
+                        :router="true" :default-active="activePath">
+                        <!-- submenu一级菜单 -->
+                        <el-submenu :index="item.path" v-for="item in menulist" :key="item.id">
+                            <!-- 一级菜单的内容渲染区域 -->
+                            <template slot="title">
+                                <!-- icon图标 -->
+                                <i :class="iconObj[item.id]"></i>
+                                <!-- 菜单文本 -->
+                                <span>{{item.authName}}</span>
+                            </template>
 
-    <!-- 脚 -->
-    <el-footer>
-        Footer
-    </el-footer>
-</el-container>
-</el-container>
-</el-container>
-</div>
+                            <!-- 二级菜单 -->
+                            <el-menu-item :index="subItem.path" v-for="subItem in item.children" :key="subItem.id" @click="saveNavState(subItem.path)">
+                                <template slot="title">
+                                    <!-- icon图标 -->
+                                    <i class="el-icon-menu"></i>
+                                    <!-- 菜单文本 -->
+                                    <span>{{subItem.authName}}</span>
+                                </template>
+                            </el-menu-item>
+                        </el-submenu>
+                    </el-menu>
+                </el-aside>
+
+                <!-- 盒子三号 -->
+                <el-container>
+                    <!-- 内容主体 -->
+                    <el-main>
+                        <router-view></router-view>
+                    </el-main>
+                    <!-- 脚 -->
+                    <el-footer>
+                        是jiojio哦
+                    </el-footer>
+                </el-container>
+            </el-container>
+        </el-container>
+    </div>
 </template>
 <script>
     export default {
@@ -65,7 +72,8 @@
             return {
                 // 左侧菜单数据
                 menulist: [],
-                //使列表 icon 循环出不一样的样式的方法
+
+                // 使列表 icon 循环出不一样的样式的方法
                 iconObj: {
                     '125': 'iconfont icon-user',
                     '103': 'iconfont icon-tijikongjian',
@@ -73,8 +81,10 @@
                     '102': 'iconfont icon-danju',
                     '145': 'iconfont icon-baobiao',
                 },
+
                 isCollapse: false,
-                //被激活的链接地址
+
+                // 被激活的链接地址
                 activePath: ''
             }
         },
@@ -83,13 +93,29 @@
             this.getMenuList();
             this.activePath = window.sessionStorage.getItem('activePath')
         },
+
         methods: {
             //退出登录
             logout() {
-                // 清空session
-                window.sessionStorage.clear();
-                // 重定向
-                this.$router.push("/login");
+                this.$confirm('即将退出登录, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+                }).then(() => {
+                    // 清空session
+                    window.sessionStorage.clear();
+                    this.$message({
+                        type: 'success',
+                        message: '成功退出'
+                    });
+                    // 重定向到登录界面
+                    this.$router.push("/login");
+                }).catch(() => {
+                    this.$message({
+                        type: 'info',
+                        message: '取消退出'
+                    });
+                });
             },
             async getMenuList() {
                 const {
@@ -112,6 +138,7 @@
         }
     }
 </script>
+
 <style lang="less" scoped>
     .home-container {
         height: 100%;
